Refactor fetch emails list by query API helper Return messages array only Improve error tracing
